pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Егор Ремезов <agrobobrov@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-json-widget = "^1.1.1"
factory-boy = "^3.3.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
force-exclude = 'migrations'

[tool.ruff]
select = ["F", "E", "W", "C90",
    "I", "N", "S", "B", "A",
    "ISC", "T20", "Q", "PTH"]
extend-exclude = [
    'migrations',
    'docker',
    'database',
    'logs,'
]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"
avoid-escape = false